pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.0"
description = ""
authors = ["ESergievich <sergievichevgeniy@gmail.com>"]
readme = "README.md"
packages = [{include = "hexlet_code"}]

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.0.7"
django-bootstrap5 = "^24.2"
django-filter = "^24.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
